Refactor chapter compatibility helpers into smaller modules
