+++ tests/test_api_voices.py
@@ -14,12 +14,16 @@ def voices_root(tmp_path, monkeypatch):
     import app.config
     import app.web
     import app.api.routers.voices
+    import app.api.routers.voices_helpers
+    import app.api.routers.voices_management
+    import app.api.routers.voices_bundles
     import app.jobs.speaker
 
     voices_dir = (tmp_path / "voices").resolve()
     monkeypatch.setattr(app.web, "VOICES_DIR", voices_dir)
     monkeypatch.setattr(app.config, "VOICES_DIR", voices_dir)
     monkeypatch.setattr(app.api.routers.voices, "VOICES_DIR", voices_dir)
+    monkeypatch.setattr(app.api.routers.voices_helpers, "VOICES_DIR", voices_dir)
     monkeypatch.setattr(app.jobs.speaker, "VOICES_DIR", voices_dir)
     return voices_dir
 
@@ -57,7 +61,7 @@ def test_list_speaker_profiles(clean_db, voices_root, client):
     (voices_dir / "SpeakerA" / "Default" / "profile.json").write_text(json.dumps({"variant_name": "Default"}))
     (voices_dir / "SpeakerA" / "Default" / "v1.wav").write_text("audio")
 
-    with patch("app.api.routers.voices.get_speaker_settings", return_value={"built_samples": [], "speed": 1.0, "test_text": "", "engine": "xtts"}):
+    with patch("app.jobs.get_speaker_settings", return_value={"built_samples": [], "speed": 1.0, "test_text": "", "engine": "xtts"}):
         response = client.get("/api/speaker-profiles")
         assert response.status_code == 200
         data = response.json()
@@ -209,7 +213,7 @@ def test_imported_latent_voice_lists_ready_without_rebuild(clean_db, voices_root
 
     bridge = MagicMock()
     bridge.check_readiness.return_value = (True, "ready")
-    with patch("app.api.routers.voices.create_voice_bridge", return_value=bridge):
+    with patch("app.engines.bridge.create_voice_bridge", return_value=bridge):
         profiles = client.get("/api/speaker-profiles").json()
 
     profile = next(p for p in profiles if p["name"] == "LatentVoice")
@@ -318,7 +322,7 @@ def test_create_character_blank_voice_uses_default(clean_db, voices_root, client
 def test_create_profile_persists_engine_metadata(clean_db, voices_root, client):
     voices_root.mkdir()
 
-    with patch("app.api.routers.voices._is_engine_active", return_value=True):
+    with patch("app.api.routers.voices_helpers._is_engine_active", return_value=True):
         response = client.post("/api/speaker-profiles", data={"speaker_id": "S1", "variant_name": "Vox", "engine": "voxtral"})
     assert response.status_code == 200
 
@@ -331,7 +335,7 @@ def test_create_profile_persists_engine_metadata(clean_db, voices_root, client):
 def test_create_voxtral_profile_requires_api_key(clean_db, voices_root, client, monkeypatch):
     voices_root.mkdir()
 
-    with patch("app.api.routers.voices._is_engine_active", return_value=False):
+    with patch("app.api.routers.voices_helpers._is_engine_active", return_value=False):
         response = client.post("/api/speaker-profiles", data={"speaker_id": "S1", "variant_name": "Vox", "engine": "voxtral"})
     assert response.status_code == 400
     assert "not enabled" in response.json()["message"]
@@ -343,7 +347,7 @@ def test_update_profile_engine(clean_db, voices_root, client):
     profile_dir.mkdir()
     (profile_dir / "profile.json").write_text(json.dumps({"variant_name": "Default"}))
 
-    with patch("app.api.routers.voices._is_engine_active", return_value=True):
+    with patch("app.api.routers.voices_helpers._is_engine_active", return_value=True):
         response = client.post("/api/speaker-profiles/SpeakerA/engine", data={"engine": "voxtral"})
     assert response.status_code == 200
     assert response.json()["engine"] == "voxtral"
@@ -358,7 +362,7 @@ def test_update_voxtral_engine_requires_api_key(clean_db, voices_root, client, m
     profile_dir.mkdir()
     (profile_dir / "profile.json").write_text(json.dumps({"variant_name": "Default", "engine": "xtts"}))
 
-    with patch("app.api.routers.voices._is_engine_active", return_value=False):
+    with patch("app.api.routers.voices_helpers._is_engine_active", return_value=False):
         response = client.post("/api/speaker-profiles/SpeakerA/engine", data={"engine": "voxtral"})
     assert response.status_code == 400
     assert "not enabled" in response.json()["message"]
@@ -381,7 +385,7 @@ def test_update_profile_reference_sample(clean_db, voices_root, client):
     (profile_dir / "profile.json").write_text(json.dumps({"variant_name": "Default", "engine": "voxtral"}))
     (profile_dir / "sample1.wav").write_text("audio")
 
-    with patch("app.api.routers.voices._is_engine_active", return_value=True):
+    with patch("app.api.routers.voices_helpers._is_engine_active", return_value=True):
         response = client.post("/api/speaker-profiles/SpeakerA/reference-sample", data={"sample_name": "sample1.wav"})
     assert response.status_code == 200
     assert response.json()["reference_sample"] == "sample1.wav"
@@ -396,7 +400,7 @@ def test_update_profile_voxtral_voice_id(clean_db, voices_root, client):
     profile_dir.mkdir()
     (profile_dir / "profile.json").write_text(json.dumps({"variant_name": "Default", "engine": "voxtral"}))
 
-    with patch("app.api.routers.voices._is_engine_active", return_value=True):
+    with patch("app.api.routers.voices_helpers._is_engine_active", return_value=True):
         response = client.post("/api/speaker-profiles/SpeakerA/voxtral-voice-id", data={"voice_id": "voice_123"})
     assert response.status_code == 200
     assert response.json()["voxtral_voice_id"] == "voice_123"
@@ -414,7 +418,7 @@ def test_voxtral_profile_test_accepts_saved_voice_id_without_samples(clean_db, v
         "voxtral_voice_id": "voice_123",
     }))
 
-    with patch("app.api.routers.voices._is_engine_active", return_value=True), patch("app.api.routers.voices.put_job"), patch("app.api.routers.voices.enqueue"):
+    with patch("app.api.routers.voices_helpers._is_engine_active", return_value=True), patch("app.state.put_job"), patch("app.jobs.enqueue"):
         response = client.post("/api/speaker-profiles/SpeakerA/test")
 
     assert response.status_code == 200
@@ -431,7 +435,7 @@ def test_voice_test_job_uses_descriptive_queue_title(clean_db, voices_root, clie
         "engine": "xtts",
     }))
 
-    with patch("app.api.routers.voices.put_job") as mock_put_job, patch("app.api.routers.voices.enqueue"):
+    with patch("app.state.put_job") as mock_put_job, patch("app.jobs.enqueue"):
         response = client.post("/api/speaker-profiles/Alice%20-%20Warm/test")
 
     assert response.status_code == 200
@@ -498,7 +502,7 @@ def test_rename_profile_and_security(clean_db, voices_root, client):
     assert response.status_code == 403
 
 def test_speaker_settings_updates(clean_db, client):
-    with patch("app.api.routers.voices.update_speaker_settings") as mock_update:
+    with patch("app.jobs.update_speaker_settings") as mock_update:
         # Test text
         response = client.post("/api/speaker-profiles/SpeakerA/test-text", data={"text": "Hello world"})
         assert response.status_code == 200
@@ -532,13 +536,13 @@ def test_build_and_test_profiles(clean_db, voices_root, client):
     voices_dir = voices_root
     voices_dir.mkdir()
 
-    with patch("app.api.routers.voices.VOICES_DIR", voices_dir):
+    with patch("app.api.routers.voices_helpers.VOICES_DIR", voices_dir):
         # Build
         file_content = b"fake wav"
         files = {"files": ("input1.wav", io.BytesIO(file_content), "audio/wav")}
         (voices_dir / "SpeakerA").mkdir()
         (voices_dir / "SpeakerA" / "profile.json").write_text("{}")
-        with patch("app.api.routers.voices.put_job"), patch("app.api.routers.voices.enqueue"):
+        with patch("app.state.put_job"), patch("app.jobs.enqueue"):
             response = client.post("/api/speaker-profiles/SpeakerA/build", files=files)
             assert response.status_code == 200
             # Build creates SpeakerA/Default for a flat name if it's missing
@@ -547,7 +551,7 @@ def test_build_and_test_profiles(clean_db, voices_root, client):
         (voices_dir / "SpeakerA" / "1.wav").write_text("fake wav content 2")
 
         # Test
-        with patch("app.api.routers.voices.put_job"), patch("app.api.routers.voices.enqueue"):
+        with patch("app.state.put_job"), patch("app.jobs.enqueue"):
             response = client.post("/api/speaker-profiles/SpeakerA/test")
             assert response.status_code == 200
 
@@ -555,11 +559,11 @@ def test_build_and_test_require_samples(clean_db, voices_root, client):
     voices_dir = voices_root
     voices_dir.mkdir()
 
-    with patch("app.api.routers.voices.VOICES_DIR", voices_dir):
+    with patch("app.api.routers.voices_helpers.VOICES_DIR", voices_dir):
         profile_dir = voices_dir / "SpeakerA"
         profile_dir.mkdir(parents=True, exist_ok=True)
 
-        with patch("app.api.routers.voices.put_job"), patch("app.api.routers.voices.enqueue"):
+        with patch("app.state.put_job"), patch("app.jobs.enqueue"):
             response = client.post("/api/speaker-profiles/SpeakerA/test")
             assert response.status_code == 400
             assert "sample" in response.json()["message"].lower()
@@ -579,7 +583,7 @@ def test_xtts_voice_actions_reject_when_disabled(clean_db, voices_root, client):
     }))
     (profile_dir / "1.wav").write_text("fake wav content")
 
-    with patch("app.api.routers.voices._is_engine_active", return_value=False):
+    with patch("app.api.routers.voices_helpers._is_engine_active", return_value=False):
         response = client.post("/api/speaker-profiles/SpeakerA/test")
 
     assert response.status_code == 400
@@ -590,7 +594,7 @@ def test_legacy_build_and_rename(clean_db, voices_root, client):
     voices_dir.mkdir()
 
     # Legacy Build
-    with patch("app.api.routers.voices.put_job"), patch("app.api.routers.voices.enqueue"):
+    with patch("app.state.put_job"), patch("app.jobs.enqueue"):
         response = client.post("/api/speaker-profiles/SpeakerL/build", files={"files": ("sample.wav", io.BytesIO(b"legacy"), "audio/wav")})
         assert response.status_code == 200
 
@@ -615,7 +619,7 @@ def test_profile_creation_errors(clean_db, voices_root, client):
     assert response.status_code == 403
 
     # Exception
-    with patch("app.api.routers.voices.Path.mkdir", side_effect=Exception("boom")):
+    with patch("app.api.routers.voices_management.Path.mkdir", side_effect=Exception("boom")):
         response = client.post("/api/speaker-profiles", data={"speaker_id": "Err", "variant_name": "V1"})
         assert response.status_code == 500
 
@@ -667,7 +671,7 @@ def test_upload_samples_security_and_failure(clean_db, voices_root, client):
     files = {"files": ("sample.wav", io.BytesIO(b"data"), "audio/wav")}
 
     # Exception during upload
-    with patch("app.api.routers.voices.Path.mkdir", side_effect=Exception("makedirs failed")):
+    with patch("app.api.routers.voices_actions.Path.mkdir", side_effect=Exception("makedirs failed")):
         response = client.post("/api/speaker-profiles/SpeakerA/samples/upload", files=files)
         assert response.status_code == 500
 
@@ -697,7 +701,7 @@ def test_delete_sample_rejects_traversal(voices_root):
     voices_dir.mkdir()
     (voices_dir / "SpeakerA").mkdir()
     (voices_dir / "SpeakerA" / "profile.json").write_text("{}")
-    with patch("app.api.routers.voices.VOICES_DIR", voices_dir):
+    with patch("app.api.routers.voices_helpers.VOICES_DIR", voices_dir):
         response = delete_speaker_sample(
             name="SpeakerA",
             sample_name="../../escape.wav",
@@ -711,6 +715,6 @@ def test_assign_profile_to_speaker_errors(clean_db, voices_root, client):
     (voices_dir / "SomeProf" / "profile.json").write_text("{}")
 
     # Generic error
-    with patch("app.api.routers.voices.get_speaker", side_effect=Exception("db crash")):
+    with patch("app.api.routers.voices_management.db.get_speaker", side_effect=Exception("db crash")):
         response = client.post("/api/speaker-profiles/SomeProf/assign", data={"speaker_id": "sid"})
         assert response.status_code == 500
